simplify mcp server class, move to actors/call.ts
